requirements.txt
numpy==1.26.4
matplotlib==3.8.4
seaborn==0.13.2
opencv-python==4.9.0.80
tensorflow-cpu==2.17.0
scikit-learn==1.4.2
Pillow==10.3.0
tqdm==4.66.4
pandas==2.2.2
streamlit==1.35.0
